requirements.txt
apex==0.9.10dev
matplotlib==3.5.1
monai==0.9.0
numpy==1.22.3
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.4.0
PyYAML==6.0
ruamel_yaml==0.17.21
scikit_image==0.19.3
scipy==1.7.3
scikit-image
tensorflow==2.12.0
timm==0.6.13
tqdm==4.64.1
git+https://github.com/huggingface/transformers
evaluate
